+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changes by Version
 
 <!-- next version -->
+## v1.29.0
+- Add `soft_limit_unique_metrics` to the [`logzio_metrics_account`](./docs/resources/metrics_account.md) resource, for Consumption accounts.
+- Add support for the [`logzio_tracing_account`](./docs/resources/tracing_account.md) resource, for Consumption accounts.
+
 ## v1.28.0
 - Upgrade `logzio_client_terraform` to `1.31.1`.
 - Add support for the [Unified Dashboards](./docs/resources/unified_dashboard.md) resource and [data source](./docs/data-sources/unified_dashboard.md).